+++ pyxel/models/charge_measurement/__init__.py
@@ -15,6 +15,11 @@ from .readout_noise import (
     readout_noise_saphira,
 )
 from .nghxrg.nghxrg import nghxrg
-from .linearity import output_node_linearity_poly
+from .linearity import (
+    output_node_linearity_poly,
+    simple_physical_non_linearity,
+    physical_non_linearity,
+    physical_non_linearity_with_saturation,
+)
 from .offset import dc_offset, output_pixel_reset_voltage_apd
 from .reset_noise import ktc_noise